FORMS-16943: updated tests
